+++ cmd/vegabenchmark/setup.go
@@ -59,7 +59,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 
 	commander := mocks.NewMockCommander(ctrl)
 	commander.EXPECT().
-		Command(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
+		Command(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
 		AnyTimes().
 		Return(nil)