+++ evtforward/forwarder.go
@@ -16,6 +16,7 @@ import (
 	"code.vegaprotocol.io/vega/metrics"
 	"code.vegaprotocol.io/vega/txn"
 
+	"github.com/cenkalti/backoff"
 	"github.com/golang/protobuf/proto"
 )
 
@@ -34,8 +35,8 @@ type TimeService interface {
 
 //go:generate go run github.com/golang/mock/mockgen -destination mocks/commander_mock.go -package mocks code.vegaprotocol.io/vega/evtforward Commander
 type Commander interface {
-	Command(ctx context.Context, cmd txn.Command, payload proto.Message, f func(error))
-	CommandSync(ctx context.Context, cmd txn.Command, payload proto.Message, f func(error))
+	Command(ctx context.Context, cmd txn.Command, payload proto.Message, f func(error), bo *backoff.ExponentialBackOff)
+	CommandSync(ctx context.Context, cmd txn.Command, payload proto.Message, f func(error), bo *backoff.ExponentialBackOff)
 }
 
 //go:generate go run github.com/golang/mock/mockgen -destination mocks/validator_topology_mock.go -package mocks code.vegaprotocol.io/vega/evtforward ValidatorTopology
@@ -279,7 +280,7 @@ func (f *Forwarder) send(ctx context.Context, evt *commandspb.ChainEvent) {
 		if err != nil {
 			f.log.Error("could not send command", logging.String("tx-id", evt.TxId), logging.Error(err))
 		}
-	})
+	}, nil)
 }
 
 func (f *Forwarder) isSender(evt *commandspb.ChainEvent) bool {