+++ monitor/price/pricemonitoring.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"sort"
+	"sync"
 	"time"
 
 	proto "code.vegaprotocol.io/protos/vega"
@@ -117,6 +118,7 @@ type Engine struct {
 
 	refPriceCacheTime time.Time
 	refPriceCache     map[int64]num.Decimal
+	refPriceLock      sync.RWMutex
 
 	boundFactorsInitialised bool
 
@@ -359,6 +361,7 @@ func (e *Engine) reset(price *num.Uint, volume uint64, now time.Time) {
 		}
 	}
 	e.priceRangeCacheTime = time.Time{}
+	e.refPriceCacheTime = time.Time{}
 	// we're not reseetting the down/up factors - they will be updated as triggered by auction end/time
 	e.reactivateBounds()
 	e.stateChanged = true
@@ -444,7 +447,7 @@ func (e *Engine) getCurrentPriceRanges(force bool) map[*bound]priceRange {
 		if !b.Active {
 			continue
 		}
-		ref := e.getRefPrice(b.Trigger.Horizon)
+		ref := e.getRefPrice(b.Trigger.Horizon, force)
 		var min, max num.Decimal
 
 		if e.boundFactorsInitialised {
@@ -493,10 +496,13 @@ func (e *Engine) clearStalePrices() {
 }
 
 // getRefPrice caches and returns the ref price for a given horizon. The cache is invalidated when block changes.
-func (e *Engine) getRefPrice(horizon int64) num.Decimal {
-	if e.refPriceCacheTime != e.now {
+func (e *Engine) getRefPrice(horizon int64, force bool) num.Decimal {
+	e.refPriceLock.Lock()
+	defer e.refPriceLock.Unlock()
+	if e.refPriceCacheTime != e.now || force {
 		e.refPriceCache = make(map[int64]num.Decimal, len(e.refPriceCache))
 		e.stateChanged = true
+		e.refPriceCacheTime = e.now
 	}
 
 	if _, ok := e.refPriceCache[horizon]; !ok {
@@ -504,6 +510,18 @@ func (e *Engine) getRefPrice(horizon int64) num.Decimal {
 		e.stateChanged = true
 	}
 	return e.refPriceCache[horizon]
+}
+
+func (e *Engine) getRefPriceNoUpdate(horizon int64) num.Decimal {
+	e.refPriceLock.RLock()
+	defer e.refPriceLock.RUnlock()
+	if e.refPriceCacheTime == e.now {
+		if _, ok := e.refPriceCache[horizon]; !ok {
+			return e.calculateRefPrice(horizon)
+		}
+		return e.refPriceCache[horizon]
+	}
+	return e.calculateRefPrice(horizon)
 }
 
 // calculateRefPrice returns theh last VolumeWeightedPrice with time preceding currentTime - horizon seconds. If there's only one price it returns the Price.